requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
alembic==1.14.0
httpx==0.28.1
email-validator==2.2.0